requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
fastapi
pydantic
uvicorn
mysql-connector-python
pymongo[srv]
redis
confluent-kafka